requirements.txt
Django==1.8.8
django-admin-bootstrapped==2.5.7
django-modeltranslation==0.10.2
google-api-python-client==1.4.2
httplib2==0.9.2
oauth2client==1.5.2
Pillow==3.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.2.3
simplejson==3.8.1
six==1.10.0
uritemplate==0.6
wheel==0.24.0
